+++ mmaction/models/recognizers/recognizer2d.py
@@ -83,3 +83,36 @@ class Recognizer2D(BaseRecognizer):
         x = self.extract_feat(imgs)
         outs = (self.cls_head(x, num_segs), )
         return outs
+
+    def forward_gradcam(self, imgs):
+        """Defines the computation performed at every call when using gradcam
+        utils."""
+        test_crops = self.test_cfg.get('test_crops', None)
+        twice_sample = self.test_cfg.get('twice_sample', False)
+
+        batches = imgs.shape[0]
+
+        imgs = imgs.reshape((-1, ) + imgs.shape[2:])
+        num_segs = imgs.shape[0] // batches
+
+        losses = dict()
+
+        x = self.extract_feat(imgs)
+        if hasattr(self, 'neck'):
+            x = [
+                each.reshape((-1, num_segs) +
+                             each.shape[1:]).transpose(1, 2).contiguous()
+                for each in x
+            ]
+            x, loss_aux = self.neck(x)
+            x = x.squeeze(2)
+            losses.update(loss_aux)
+            num_segs = 1
+
+        cls_score = self.cls_head(x, num_segs)
+        if test_crops is not None:
+            if twice_sample:
+                test_crops = test_crops * 2
+            cls_score = self.average_clip(cls_score, test_crops)
+
+        return cls_score